Added Propellor.Property.File.configFileName and related functions to generate good filenames for config directories.
spwhitton has a branch that could use this, and there are several places
in propellor that do something ad-hoc that would have been better
implemented using this. I was not able to switch any of the existing
ad-hoc stuff, but this can be used going forward for new stuff.

This commit was sponsored by Anthony DeRobertis on Patreon.